[mob] Refactor local import
