+++ mobile/lib/services/sync/import/model.dart
@@ -1,0 +1,11 @@
+class LocalPathAsset {
+  final Set<String> localIDs;
+  final String pathID;
+  final String pathName;
+
+  LocalPathAsset({
+    required this.localIDs,
+    required this.pathName,
+    required this.pathID,
+  });
+}
